pyproject.toml
[project]
name = "medicare-portal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi>=0.115.6",
    "httpx>=0.28.1",
    "jinja2>=3.1.4",
    "libsql-experimental>=0.0.41",
    "passlib>=1.7.4",
    "pydantic>=2.10.3",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.20",
    "python-slugify>=8.0.4",
    "sendgrid>=6.11.0",
    "uvicorn>=0.32.1",
]
